POSIX2 work
Added incomplete support for libraries depending on other internal
libraries. The extra dependency is not yet detected by the final target.

Fixed module.mk.in for c-util and associated bugs in gencache

Various bugfixes.

+++ src/zoidfs/dispatcher/posix2/dcache.c
@@ -64,11 +64,11 @@ static int dcache_key_compare (const void * i1, const void * i2)
          sizeof (((zoidfs_handle_t *)i1)->data)) == 0 ? 1 : 0); 
 }
 
-static int dcache_key_hash (gencache_key_t key)
+static unsigned long dcache_key_hash (gencache_key_t key)
 {
    const zoidfs_handle_t * h = (const zoidfs_handle_t *) key; 
    const int * ptr = (int *) &h->data; 
-   int init = 66; 
+   unsigned long init = 66; 
    unsigned int i; 
    assert (sizeof (h->data) % sizeof(int) == 0); 
    for (i=0; i<sizeof (h->data)/sizeof(int); ++i)
